# apis/app-api/order/order-app-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Order App API",
    "version": "1.0.0",
    "description": "App/client contract for sdkwork-order capability (split from retired sdkwork-commerce monolith).",
    "x-sdkwork-api-authority": "sdkwork-order-app-api",
    "x-sdkwork-sdk-family": "sdkwork-order-app-sdk",
    "x-sdkwork-audience": "App, desktop, mobile, H5, and user-facing clients.",
    "x-sdkwork-owner": "sdkwork-order"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18079",
      "description": "Local sdkwork-order via IM gateway"
    }
  ],
...
  "paths": {
...
    "/app/v3/api/memberships/orders": {
      "post": {
        "tags": [
          "orderMemberships"
        ],
        "summary": "Create or reuse a membership purchase-intent order.",
        "operationId": "memberships.orders.create",
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipOrderCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-order",
        "x-sdkwork-api-authority": "sdkwork-order-app-api",
        "x-sdkwork-domain": "commerce",
        "x-sdkwork-resource": "memberships.orders",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "parameters": [
          {
            "$ref": "#/components/parameters/WriteCommandIdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MembershipOrderCreateCommand"
              }
            }
          }
        },
        "x-sdkwork-idempotent": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "SDKWork auth token carried as Authorization: Bearer <auth_token>."
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "SDKWork access isolation token."
      }
    },
    "schemas": {
...
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40011,
            "maximum": 40099
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
...
      "MembershipOrderCreateCommand": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "action",
          "packageId",
          "paymentMethod"
        ],
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "purchase",
              "renew",
              "upgrade",
              "recharge"
            ],
            "description": "Membership lifecycle action. Recharge adds quota to the current active subscription's entitlement account; it participates in business purchase-intent identity."
          },
          "packageId": {
            "type": "string",
            "description": "Catalog package external id (purchase/renew/upgrade). Recharge uses the synthetic 'membership-quota-recharge' package id and ignores the catalog."
          },
          "paymentMethod": {
            "type": "string"
          },
          "paymentProduct": {
            "type": "string",
            "enum": [
              "mobile_cashier_h5",
              "wechat_native",
              "alipay_native"
            ],
            "default": "mobile_cashier_h5",
            "description": "QR payment product. H5 returns the order-bound cashierUrl; native products create a provider payment intent."
          },
          "clientRequestNo": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "grantQuantity": {
            "type": "integer",
            "format": "int64",
            "description": "Quota units to add to the membership entitlement account (required for action=recharge)."
          },
          "amount": {
            "type": "string",
            "description": "Recharge price in currency amount (required for action=recharge)."
          }
        }
      },
      "MembershipOrderCreateResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "item"
                ],
                "properties": {
                  "item": {
                    "$ref": "#/components/schemas/MembershipOrderCreateResult"
                  }
                }
              }
            }
          }
        ]
      },
      "MembershipOrderCreateResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "action",
          "orderId",
          "orderNo",
          "outTradeNo",
          "amount",
          "currencyCode",
          "packageId",
          "packageName",
          "durationDays",
          "expiresAt",
          "paymentMethod",
          "paymentProduct",
          "qrCode",
          "qrCodeType",
          "paymentParams",
          "status",
          "reused",
          "cashierUrl",
          "grantQuantity"
        ],
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "purchase",
              "renew",
              "upgrade",
              "recharge"
            ]
          },
          "orderId": {
            "type": "string"
          },
          "orderNo": {
            "type": "string"
          },
          "outTradeNo": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "currencyCode": {
            "type": "string"
          },
          "packageId": {
            "type": "string"
          },
          "packageName": {
            "type": "string"
          },
          "durationDays": {
            "type": "integer",
            "format": "int64"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "paymentMethod": {
            "type": "string"
          },
          "paymentProduct": {
            "type": "string",
            "enum": [
              "mobile_cashier_h5",
              "wechat_native",
              "alipay_native"
            ]
          },
          "qrCode": {
            "type": "string"
          },
          "qrCodeType": {
            "type": "string",
            "enum": [
              "cashier_url",
              "provider_native"
            ]
          },
          "paymentId": {
            "type": [
              "string",
              "null"
            ]
          },
          "paymentParams": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "status": {
            "type": "string"
          },
          "reused": {
            "type": "boolean",
            "description": "True when an idempotency replay or an existing active purchase intent was returned."
          },
          "cashierUrl": {
            "type": "string"
          },
          "grantQuantity": {
            "type": "integer",
            "format": "int64",
            "description": "Quota units granted by a membership quota recharge order (present when action=recharge)."
          }
        }
      },
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
...
    },
    "parameters": {
      "WriteCommandIdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1,
          "maxLength": 128
        },
        "description": "Client retry key scoped by tenant, principal, method, and path."
      }
    }
  },
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ]
}